hypervisor: move inner class, add property to choose libvirt protocol, live migration with tunneled and peer-2-peer flags on.
